refactor(test): group security coverage
