+++ test/security/config-set-nested-ssrf.test.ts
@@ -23,9 +23,9 @@ function installMockPrivilegedExec(
   events?: string[],
 ): MockGuardRestore {
   const priorPrivilegedExec = require.cache[privilegedExecPath];
-  const timerBoundLockPath = require.resolve("../src/lib/shields/timer-bound-lock");
+  const timerBoundLockPath = require.resolve("../../src/lib/shields/timer-bound-lock");
   const priorTimerBoundLock = require.cache[timerBoundLockPath];
-  const openClawConfigLockPath = require.resolve("../src/lib/shields/openclaw-config-lock");
+  const openClawConfigLockPath = require.resolve("../../src/lib/shields/openclaw-config-lock");
   const priorOpenClawConfigLock = require.cache[openClawConfigLockPath];
   requireCache[privilegedExecPath] = {
     id: privilegedExecPath,
@@ -110,10 +110,10 @@ function installMockPrivilegedExec(
 
 describe("config set nested URL SSRF enforcement", () => {
   it("rejects nested object/array URL values that target private hosts", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -156,7 +156,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
       const nestedValue = JSON.stringify({
         primary: "https://api.nvidia.com/v1",
         fallback: ["https://example.com/v1", { internal: "http://localhost:8080/internal" }],
@@ -191,10 +191,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("validates the key before doing URL or DNS validation", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -241,7 +241,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
 
       await expect(
         configSet("sandbox-ssrf-test", {
@@ -274,10 +274,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("accepts nested object/array URL values when all are public", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -322,7 +322,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
       const nestedValue = JSON.stringify({
         primary: "https://93.184.216.34/v1",
         fallback: ["http://93.184.216.35/v1", { backup: "https://93.184.216.36/v2" }],
@@ -376,10 +376,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("writes an OpenShell bridge URL in an OpenClaw provider baseUrl (#7453)", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -420,7 +420,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
       const provider = {
         api: "ollama",
         baseUrl: "http://host.openshell.internal:11434",
@@ -460,10 +460,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("surfaces an OpenClaw schema rejection as clean SandboxConfigError lines", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
     const priorShieldsAudit = require.cache[shieldsAuditPath];
@@ -498,7 +498,7 @@ describe("config set nested URL SSRF enforcement", () => {
     } as any;
 
     try {
-      const { configSet, SandboxConfigError } = require("../src/lib/sandbox/config");
+      const { configSet, SandboxConfigError } = require("../../src/lib/sandbox/config");
       let thrown: unknown;
       try {
         await configSet("sandbox-schema-test", {
@@ -527,10 +527,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("ignores nested non-http URL-like strings and does not crash", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -573,7 +573,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
       const nestedValue = JSON.stringify({
         ftpUrl: "ftp://files.example.com",
         plainText: "not-a-url",
@@ -613,10 +613,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("recognizes mixed-case http and https schemes in nested values", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -659,7 +659,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
       const nestedValue = JSON.stringify({
         primary: "HTTP://93.184.216.34/v1",
         fallback: ["HtTpS://93.184.216.35/v2", { backup: "hTtP://93.184.216.36/v3" }],
@@ -698,10 +698,10 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("redacts credentials, query strings, and fragments in validation errors", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const privilegedExecPath = require.resolve("../src/lib/sandbox/privileged-exec");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const privilegedExecPath = require.resolve("../../src/lib/sandbox/privileged-exec");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -744,7 +744,7 @@ describe("config set nested URL SSRF enforcement", () => {
     const logSpy = vi.spyOn(console, "log").mockImplementation(() => {});
 
     try {
-      const { configSet } = require("../src/lib/sandbox/config");
+      const { configSet } = require("../../src/lib/sandbox/config");
       const nestedValue = JSON.stringify({
         primary: "http://user:pass@127.0.0.1:8080/private/path?token=secret#frag",
       });
@@ -791,11 +791,11 @@ describe("config set nested URL SSRF enforcement", () => {
   });
 
   it("records the config rotate-token audit action as rotate_token (not shields_down)", async () => {
-    const sandboxConfigPath = require.resolve("../src/lib/sandbox/config");
-    const openshellPath = require.resolve("../src/lib/adapters/openshell/client");
-    const shieldsAuditPath = require.resolve("../src/lib/shields/audit");
-    const sessionPath = require.resolve("../src/lib/state/onboard-session");
-    const credStorePath = require.resolve("../src/lib/credentials/store");
+    const sandboxConfigPath = require.resolve("../../src/lib/sandbox/config");
+    const openshellPath = require.resolve("../../src/lib/adapters/openshell/client");
+    const shieldsAuditPath = require.resolve("../../src/lib/shields/audit");
+    const sessionPath = require.resolve("../../src/lib/state/onboard-session");
+    const credStorePath = require.resolve("../../src/lib/credentials/store");
 
     const priorSandboxConfig = require.cache[sandboxConfigPath];
     const priorOpenshell = require.cache[openshellPath];
@@ -858,7 +858,7 @@ describe("config set nested URL SSRF enforcement", () => {
     process.env.ROTATE_NEW_TOKEN = "nvapi-rotated-value";
 
     try {
-      const { configRotateToken } = require("../src/lib/sandbox/config");
+      const { configRotateToken } = require("../../src/lib/sandbox/config");
 
       await expect(
         configRotateToken("rotate-test", { fromEnv: "ROTATE_NEW_TOKEN" }),
